# fantastic-coffee-decaffeinated/doc/api.yaml
openapi: 3.0.1
info:
  title: WASAPhoto
  description: |
    This API allows to manage an instagram-like web application
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /profiles/{username}:
    parameters:
      - $ref: "#/components/parameters/username"
    get:
      tags: ["user"]
      operationId: getUserProfile
      summary: Return one user profile
      description: |
        Get the user profile with associated username
      responses:
        "200":
          description: Download the specified user profile page
          content:
            application/json:
              schema:
                description: |
                  a thumbnail object list with photos and informations reelated to them
                type: object
                properties:
                  thumbnails:
                    description: a thumbnail object with the photo and informations
                    type: array
                    items:
                      $ref: "#/components/schemas/thumbnailObject"
                    minItems: 0
                    maxItems: 1000
        "400":
          description: |
            the username provided for the user profile is not valid
          content:
           application/json:
            schema:
              $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: User not found
          content:
           application/json:
            schema:
              $ref: "#/components/schemas/generalError"
        '500':
          description: The server cannot elaborate the request to get the profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
    put:
      tags: ["user"]
      operationId: setMyUserName
      summary: set the user name
      description: Allow to set a new username
      requestBody:
        description: the new username to set
        content:
          application/json:
            schema:
              description: The new username
              type: object
              properties:
                name:
                  description: The new user's username to set
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
            examples:
              name:
                summary: maria-user
                value:
                  name: maria
        required: true
      responses:
        "200":
          description: Username updated, everything fine
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: The username is not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: The server cannot elaborate the username update request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
...
  /photos:
    post:
      tags: ["photo"]
      operationId: uploadPhoto
      summary: Upload a user photo
      description: |
        Upload a photo from the current user logged in
      requestBody:
        content:
          image/*:
            schema:
              description: The photo file base64 rappresentation
              type: string
              format: base64
      responses:
        "201":
          description: Photo uploaded
          content:
            application/json:
              schema:
                description: A json object with one property containing the photo id
                type: object
                properties:
                  idphoto:
                    description: The photo id
                    type: string
              examples:
                photouser:
                  summary: maria-photo
                  value:
                    idphoto: abcd20132303151010
        "400":
          description: Photo format/photo id/user id not valid for uploadPhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: The server cannot elaborate deconding photo file request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
...
  /follows/{username}:
    parameters:
      - $ref: "#/components/parameters/username"
    delete:
      tags: ["user"]
      operationId: unfollowUser
      summary: Unfollow one user
      description: |
        Allow to unfollow on user
      responses:
        "200":
          description: User unfollowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: Invalid user id unfollowUser
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "500":
          description: The server cannot proces the unfollowUser request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /bans/{username}:
    parameters:
      - $ref: "#/components/parameters/username" 
    delete:
      tags: ["user"]
      operationId: unbanUser  
      summary: unban a user
      description: |
        Allow to unban the username as path parameter
      responses:
        "200":
          description: user unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: user id not valid for unBanUser operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    username:
      name: username
      in: path
      required: true
      description: A username
      schema:
        type: string
...
  schemas:
    thumbnailObject:
      description: |
        A rappresentation of a thumnail pic with other informations
      type: object
      properties:
        idphoto:
          description: the photo url
          type: string
        datetime:
          description: the post date time to the referred photo
          type: string
          format: date-time
        nlikes:
          description: The likes number to the referred photo
          type: integer
        ncomments:
          description: The comments number to the referred photo
          type: integer
      example:
        photo1:
          idphoto: idphoto1234
          datetime: "2022-02-02T13:30:15"
          nlikes: 10
          ncomments: 3
        photo2:
          idphoto: idphoto4321
          datetime: "2022-03-09T17:20:24"
          nlikes: 5
          ncomments: 2
...
    generalError:
      description: An object describing an error
      type: object
      properties:
        error:
          description: The message content of the error
          type: string
          example: "error description"
    generalFeedback:
      description: The feedback provided from a response
      type: object
      properties:
        feedback:
          description: The content of the feedback
          type: string
          example: "feedback description"
  responses:
    UnauthorizedError:
      description: Access identifier is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/generalError"
security:
 - bearerAuth: []